requirements.txt
ultralytics>=8.0.0
opencv-python>=4.8.0
twilio>=8.0.0
python-dotenv>=1.0.0
torch>=2.0.0       # Apple Silicon usará MPS automáticamente
